pyproject.toml
[project]
name = "RecommandationsCyber"
version = "0.1.0"
description = "Une interface permettant d'interroger [Albert](https://albert.etalab.gouv.fr), le modèle IA, chargé avec des guides de l'ANSSI."
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "gradio",
    "httpx",
    "openai",
    "psycopg2-binary",
    "pytest",
    "requests",
    "typing",
    "uvicorn",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.ruff]
target-version = "py313"